+++ src/main/java/org/websoso/WSSServer/controller/UserController.java
@@ -47,7 +47,7 @@ import org.websoso.WSSServer.dto.userNovel.UserNovelAndNovelsGetResponseLegacy;
 import org.websoso.WSSServer.dto.userNovel.UserTasteAttractivePointPreferencesAndKeywordsGetResponse;
 import org.websoso.WSSServer.feed.service.FeedService;
 import org.websoso.WSSServer.library.service.UserNovelService;
-import org.websoso.WSSServer.service.UserService;
+import org.websoso.WSSServer.user.service.UserService;
 import org.websoso.WSSServer.validation.NicknameConstraint;
 
 @RestController
