+++ src/main/java/org/websoso/WSSServer/service/BlockService.java
@@ -16,6 +16,7 @@ import org.websoso.WSSServer.dto.block.BlockGetResponse;
 import org.websoso.WSSServer.dto.block.BlocksGetResponse;
 import org.websoso.WSSServer.exception.exception.CustomBlockException;
 import org.websoso.WSSServer.repository.BlockRepository;
+import org.websoso.WSSServer.user.service.UserService;
 
 @Service
 @RequiredArgsConstructor
