+++ src/main/java/org/websoso/WSSServer/user/service/AuthService.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.service;
+package org.websoso.WSSServer.user.service;
 
 import static org.websoso.WSSServer.exception.error.CustomAuthError.INVALID_TOKEN;
